web/worker/requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chainer==7.7.0
chainercv==0.13.1
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1.2
django-captcha==1.0.0
django-simple-email-confirmation==0.70
filelock==3.0.12
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.3.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
Pillow==8.0.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rq==1.5.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
sentry-sdk==0.19.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
threadpoolctl==2.1.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.2
